+++ src/realm/object-store/sync/sync_manager.hpp
@@ -182,12 +182,6 @@ public:
     // Removes a user
     void remove_user(const std::string& user_id) REQUIRES(!m_user_mutex, !m_file_system_mutex);
 
-    // Get the default path for a Realm for the given user and absolute unresolved URL.
-    // If the default path of `<rootDir>/<appId>/<userId>/<realm_file_name>.realm` cannot
-    // be created, this function may pass back `<rootDir>/<hashedFileName>.realm`
-    std::string path_for_realm(const SyncUser& user, const std::string& realm_file_name) const
-        REQUIRES(!m_file_system_mutex);
-
     // Get the default path for a Realm for the given configuration.
     // The default value is `<rootDir>/<appId>/<userId>/<partitionValue>.realm`.
     // If the file cannot be created at this location, for example due to path length restrictions,